+++ services/containerinstance/mgmt/2021-03-01/containerinstance/CHANGELOG.md
@@ -1,2 +1,23 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Structs
+
+1. HTTPHeaders
+
+### Signature Changes
+
+#### Struct Fields
+
+1. ContainerHTTPGet.HTTPHeaders changed type from *HTTPHeaders to *[]HTTPHeader
+
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. HTTPHeader